backend/core_service/requirements.txt
# Requirements для Django Core Service
# Электронный журнал производства работ - Промышленное решение

# Django Framework и основные компоненты
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# База данных и ORM
psycopg2-binary==2.9.9
django-postgres-extra==2.0.8

# GIS поддержка
gdal==3.8.1

# Кеширование и очереди
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-redis==5.4.0

# Аутентификация и безопасность
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-guardian==2.4.0
django-ratelimit==4.1.0
cryptography==41.0.8

# Валидация и сериализация
marshmallow==3.20.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Обработка файлов и изображений
Pillow==10.2.0
python-magic==0.4.27
django-storages==1.14.2

# Мониторинг и логирование
django-prometheus==2.3.1
structlog==23.2.0
django-structlog==8.0.0
sentry-sdk==1.40.6

# HTTP клиенты и API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Работа с датами и временем
python-dateutil==2.8.2
pytz==2023.3
django-timezone-field==6.1.0

# Конфигурация и переменные окружения
python-decouple==3.8
django-environ==0.11.2

# Валидация данных
jsonschema==4.20.0
cerberus==1.3.5

# Работа с Excel и CSV
openpyxl==3.1.2
pandas==2.1.4
xlsxwriter==3.1.9

# Работа с PDF
reportlab==4.1.0
weasyprint==60.2
python-docx==1.1.0

# Уведомления
django-notifications-hq==1.8.3

# Задачи и планировщик
APScheduler==3.10.4

# Метрики и аналитика
prometheus-client==0.19.0

# WebSocket поддержка
channels==4.0.0
channels-redis==4.1.0

# Тестирование
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.0.0
factory-boy==3.3.0
faker==23.2.1

# Утилиты разработки
ipython==8.18.1
django-debug-toolbar==4.2.0
black==24.2.0
flake8==7.0.0
isort==5.13.2

# WSGI/ASGI серверы
gunicorn==21.2.0
uvicorn==0.24.0
whitenoise==6.6.0

# AWS/Cloud хранилище
boto3==1.34.51

# Дополнительные утилиты
python-slugify==8.0.1
django-model-utils==4.3.1
django-import-export==3.3.1
